Primative Sound 1 waveform generation\n

+++ src/apu_data.h
@@ -22,7 +22,7 @@ struct apu_data
 		u16 raw_frequency;
 		double output_frequency;
 
-		u8 duration;
+		u32 duration;
 		u32 volume;
 
 		bool playing;
